chores: 주석 및 필요 없는 함수 삭제
